Refactor of cache service
Makes the following changes:

This is to help ensure the requests don't get backed up

This better reflects our requirements. We want to independently store
data for each account. We were using a map anyway, this way we et the
map handle the concurrency instead of using an atomic reference to an
immutable map.

Switches to a more event-driven model. This makes it a little easier
to reason about the execution and means we can write cache data for
each account separately, rather than threading them all together and
updating when they all complete.
This also reduces the likehood that requests will get backed up.